pyproject.toml
# App: AWS Customer CRUD
# Package: project_root
# File: pyproject.toml
# Version: 0.0.2
# Author: Bobwares
# Date: Thu Jun 05 17:10:34 UTC 2025
# Description: Project configuration.

[tool.poetry]
name = "aws-customer-crud"
version = "0.0.2"
description = "AWS Lambda CRUD application"
authors = ["Bobwares <bobwares@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = ">=1.26.0"
pydantic = ">=1.10.2"
aws-lambda-powertools = "^1.27.0"
python-jose = ">=3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
